requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
pandas==2.1.4
matplotlib==3.8.2
python-dotenv==1.0.0